OSIS-457 Moved methot that reinitialize data before proposal in business module
