+++ base/views/learning_unit_proposal.py
@@ -31,15 +31,12 @@ from django.core.exceptions import PermissionDenied
 
 
 from base.business.learning_unit_proposal import compute_form_initial_data, compute_proposal_type, \
-    is_eligible_for_modification_proposal, is_eligible_for_cancel_of_proposal
+    is_eligible_for_modification_proposal, is_eligible_for_cancel_of_proposal, reinitialize_data_before_proposal
 from base.forms.learning_unit_proposal import LearningUnitProposalModificationForm
-from base.models.entity_container_year import find_by_learning_container_year_and_linktype, EntityContainerYear
 from base.models.enums import proposal_state
 from base.models.learning_unit_year import LearningUnitYear
 from base.models.person import Person
 from base.models.proposal_learning_unit import ProposalLearningUnit
-from base.models import campus, entity
-from reference.models import language
 
 
 @login_required
@@ -87,38 +84,3 @@ def cancel_proposal_of_learning_unit(request, learning_unit_year_id):
     return redirect('learning_unit', learning_unit_year_id=learning_unit_year.id)
 
 
-def reinitialize_data_before_proposal(learning_unit_proposal, learning_unit_year):
-    initial_data = learning_unit_proposal.initial_data
-    _reinitialize_model_before_proposal(learning_unit_year, initial_data["learning_unit_year"])
-    _reinitialize_model_before_proposal(learning_unit_year.learning_unit, initial_data["learning_unit"])
-    _reinitialize_model_before_proposal(learning_unit_year.learning_container_year,
-                                        initial_data["learning_container_year"])
-    _reinitialize_entities_before_proposal(learning_unit_year.learning_container_year,
-                                           initial_data["entities"])
-
-
-def _reinitialize_model_before_proposal(obj_model, attribute_initial_values):
-    for key, value in attribute_initial_values.items():
-        if key == "id":
-            continue
-        elif key == "campus":
-            setattr(obj_model, key, campus.find_by_id(value))
-        elif key == "language":
-            setattr(obj_model, key, language.find_by_id(value))
-        else:
-            setattr(obj_model, key, value)
-    obj_model.save()
-
-
-def _reinitialize_entities_before_proposal(learning_container_year, initial_entities_by_type):
-    for type_entity, id_entity in initial_entities_by_type.items():
-        if id_entity:
-            initial_entity = entity.get_by_internal_id(id_entity)
-            if initial_entity:
-                EntityContainerYear.objects.update_or_create(learning_container_year=learning_container_year,
-                                                             type=type_entity, defaults={"entity": initial_entity})
-        else:
-            current_entity_container_year = find_by_learning_container_year_and_linktype(learning_container_year,
-                                                                                         type_entity)
-            if current_entity_container_year is not None:
-                current_entity_container_year.delete()
